// src/types/VTXChannel.ts
export enum Channel {
    R1 = 'R1',
    R2 = 'R2',
    R3 = 'R3',
    R4 = 'R4',
    R5 = 'R5',
    R6 = 'R6',
    R7 = 'R7',
    R8 = 'R8',

    A1 = 'A1',
    A2 = 'A2',
    A3 = 'A3',
    A4 = 'A4',
    A5 = 'A5',
    A6 = 'A6',
    A7 = 'A7',
    A8 = 'A8',

    B1 = 'B1',
    B2 = 'B2',
    B3 = 'B3',
    B4 = 'B4',
    B5 = 'B5',
    B6 = 'B6',
    B7 = 'B7',
    B8 = 'B8',

    E1 = 'E1',
    E2 = 'E2',
    E3 = 'E3',
    E4 = 'E4',
    E5 = 'E5',
    E6 = 'E6',
    E7 = 'E7',
    E8 = 'E8',

    F1 = 'F1',
    F2 = 'F2',
    F3 = 'F3',
    F4 = 'F4',
    F5 = 'F5',
    F6 = 'F6',
    F7 = 'F7',
    F8 = 'F8'
}

export const ChannelFrequencies: Record<Channel, number> = {
    [Channel.R1]: 5658,
    [Channel.R2]: 5695,
    [Channel.R3]: 5732,
    [Channel.R4]: 5769,
    [Channel.R5]: 5806,
    [Channel.R6]: 5843,
    [Channel.R7]: 5880,
    [Channel.R8]: 5917,

    [Channel.A1]: 5865,
    [Channel.A2]: 5845,
    [Channel.A3]: 5825,
    [Channel.A4]: 5805,
    [Channel.A5]: 5785,
    [Channel.A6]: 5765,
    [Channel.A7]: 5745,
    [Channel.A8]: 5725,

    [Channel.B1]: 5733,
    [Channel.B2]: 5752,
    [Channel.B3]: 5771,
    [Channel.B4]: 5790,
    [Channel.B5]: 5809,
    [Channel.B6]: 5828,
    [Channel.B7]: 5847,
    [Channel.B8]: 5866,

    [Channel.E1]: 5705,
    [Channel.E2]: 5685,
    [Channel.E3]: 5665,
    [Channel.E4]: 5645,
    [Channel.E5]: 5885,
    [Channel.E6]: 5905,
    [Channel.E7]: 5925,
    [Channel.E8]: 5945,

    [Channel.F1]: 5740,
    [Channel.F2]: 5760,
    [Channel.F3]: 5780,
    [Channel.F4]: 5800,
    [Channel.F5]: 5820,
    [Channel.F6]: 5840,
    [Channel.F7]: 5860,
    [Channel.F8]: 5880
};